pyproject.toml
[project]
name = "diffink-infer"
version = "0.1.0"
description = "DiffInk inference — Glyph- and Style-Aware Latent Diffusion Transformer for online handwriting generation"
requires-python = ">=3.10"
dependencies = [
    "torch==2.0.0+cu118",
    "torchvision==0.15.1+cu118",
    "torchaudio==2.0.1+cu118",
    "numpy>=1.24,<2.0",
    "h5py>=3.11",
    "einops>=0.8",
    "einx>=0.3",
    "x-transformers>=1.42",
    "pyyaml>=6.0",
    "tqdm>=4.67",
    "opencv-python>=4.11",
    "matplotlib>=3.7",
    "scipy>=1.10",
    "gdown>=5.0",
]

[project.scripts]
diffink-infer = "diffink.infer:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["diffink"]

[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu118" }
torchvision = { index = "pytorch-cu118" }
torchaudio = { index = "pytorch-cu118" }
